Add a pip requirements file for production
SSL errors leading to a server 500 error with using Google+
authentication seem have arise from the certifi package brought in by
(probably) tornado that is part of the jupyter install. All of those
packages are not necessary in production. I have therefore created a
slimmed down requirements for produciton only.