Fix issue-833 and issue-834.
For issue-834, add redirection stack push to redirect after login.
After login, if there is a 'next' key in request.GET, push that to the
redirection stack. That should be sufficient to redirect after the
subject initialization.

For issue-833, The feedback template now provides a button and also the
date completed.  Also provided a template for when feedback for an
incomplete experiment is asked for.

As a bonus, add a template for feedback for incomplete experiments.